DerpLauncher: Do not hide keyboard on clearing search query
If the user hits backspace when there is only one character left,
search gets reset back into all apps view and they keyboard gets
hidden. Then the user has to press the search bar again to open the
keyboard which is poor UX. Instead do not hide the keyboard when the
search bar is reset. Touching elsewhere on the screen already hides
the keyboard.

Change-Id: Ia41f803d6ffe258923e90ed6943aa214b193e393